requirements.txt
django==4.2.16
djangorestframework==3.15.2
requests==2.32.3
django-oauth-toolkit==2.4.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
django-environ==0.11.2
drf-yasg==1.21.7
gunicorn==22.0.0
psycopg2-binary==2.9.7
google-cloud-logging==3.10.0
whitenoise==6.7.0
django-silk==5.2.0
python-telegram-bot==21.7